requirements.txt
absl-py==0.7.0
astor==0.7.1
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
imutils==0.5.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mock==2.0.0
numpy==1.16.0
opencv-python==3.4.5.20
package-name==0.1
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
pycrypto==2.6.1
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.7.5
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
ssh==1.8.0
tensorboard==1.12.2
tensorflow==1.13.0rc1
tensorflow-estimator==1.13.0rc0
termcolor==1.1.0
urllib3==1.24.1
virtualenv==16.3.0
Werkzeug==0.14.1
wrapt==1.11.1
flask-cors
